Fix Services, controllers, new response

+++ src/main/java/com/Tisj/api/controllers/ArticuloController.java
@@ -35,33 +35,57 @@ public class ArticuloController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Articulo> getArticuloById(@PathVariable Long id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth.getAuthorities().stream()
+                .anyMatch(p -> p.getAuthority().equals("ADMIN"))) {
         Articulo articulo = articuloService.getArticuloById(id);
         if (articulo != null) {
             return new ResponseEntity<>(articulo, HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     @PostMapping
     public ResponseEntity<Articulo> createArticulo(@RequestBody Articulo articulo) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth.getAuthorities().stream()
+                .anyMatch(p -> p.getAuthority().equals("ADMIN"))) {
         Articulo nuevoArticulo = articuloService.createArticulo(articulo);
         return new ResponseEntity<>(nuevoArticulo, HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<Articulo> updateArticulo(@PathVariable Long id, @RequestBody Articulo articulo) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth.getAuthorities().stream()
+                .anyMatch(p -> p.getAuthority().equals("ADMIN"))) {
         Articulo articuloActualizado = articuloService.updateArticulo(id, articulo);
         if (articuloActualizado != null) {
             return new ResponseEntity<>(articuloActualizado, HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteArticulo(@PathVariable Long id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth.getAuthorities().stream()
+                .anyMatch(p -> p.getAuthority().equals("ADMIN"))) {
         articuloService.deleteArticulo(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 }
